Make GraphicElement::setPixmap faster (see description)
The previous implementation creates a QImage and then converts it into a QPixmap. According to the [Qt documentation](https://doc.qt.io/qt-5/qpixmap.html#pixmap-conversion), this conversion is not a fast operation.

I've resorted to defining a QPixmap and loading into it directly, which I'm assuming is faster. As the method in question is very often used, I feel like this reduction in latency is useful.