Redo the clock logic to improve performance
The first version was kept in sync but didn't work for all frequencies.
The second version set a different timepoint for each clock so they desync.
The third version incremented a fixed value inside each clock so they kept in sync but that was slow.
This version follows the logic of the second but using a global time.